Layer system rearanged
From now on there is three layers : >Presentation Layer
>Service Layer
>Persistance Layer

Presentation Layer consists of UI elements and GUI facades. It is main goal is contacting with the user and providing usefull tools to retrieve his/her desires.

Service Layer is the layer where all of these information coming from presentation layer is converted to actions, such as searching or sorting entries.

Persistance Layer is resposible for saving and serving data to the application when it is needed. It consists of three parts; database, config and encryption.